# server/ui/api/v1/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "OpenAPI 2.0 documentation for [MessagePit](https://github.com/coreydaley/messagepit).",
    "title": "MessagePit API",
    "contact": {
      "name": "GitHub",
      "url": "https://github.com/coreydaley/messagepit"
    },
    "license": {
      "name": "MIT license",
      "url": "https://github.com/coreydaley/messagepit/blob/main/LICENSE"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/sms/message/{id}": {
      "get": {
        "description": "Returns a single SMS message.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "sms"
        ],
        "summary": "Get SMS message",
        "operationId": "GetSMSMessageParams",
        "parameters": [
          {
            "type": "string",
            "description": "SMS message database ID",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/SMSMessageResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      },
      "delete": {
        "description": "Deletes a single SMS message.",
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "sms"
        ],
        "summary": "Delete SMS message",
        "operationId": "DeleteSMSMessageParams",
        "parameters": [
          {
            "type": "string",
            "description": "SMS message database ID",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/OKResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
    "/api/v1/sms/message/{id}/read": {
      "put": {
        "description": "Marks an SMS message as read.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "sms"
        ],
        "summary": "Mark SMS message read",
        "operationId": "MarkSMSReadParams",
        "parameters": [
          {
            "type": "string",
            "description": "SMS message database ID",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/OKResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
    "/api/v1/sms/messages": {
      "get": {
        "description": "Returns SMS messages ordered from newest to oldest.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "sms"
        ],
        "summary": "List SMS messages",
        "operationId": "GetSMSMessagesParams",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "default": 0,
            "x-go-name": "Start",
            "description": "Pagination offset",
            "name": "start",
            "in": "query"
          },
          {
            "type": "integer",
            "format": "int64",
            "default": 50,
            "x-go-name": "Limit",
            "description": "Limit number of results",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/SMSMessagesSummaryResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "description": "Deletes all SMS messages.",
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "sms"
        ],
        "summary": "Delete all SMS messages",
        "operationId": "DeleteAllSMSParams",
        "responses": {
          "200": {
            "$ref": "#/responses/OKResponse"
          }
        }
      }
    },
    "/api/v1/tags": {
      "get": {
        "description": "Returns a JSON array of all unique message tags.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tags"
        ],
        "summary": "Get all current tags",
        "operationId": "GetAllTags",
        "responses": {
          "200": {
            "$ref": "#/responses/ArrayResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "put": {
        "description": "This will overwrite any existing tags for selected message database IDs. To remove all tags from a message, pass an empty tags array.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tags"
        ],
        "summary": "Set message tags",
        "operationId": "SetTagsParams",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "Tags",
                "IDs"
              ],
              "properties": {
                "IDs": {
                  "description": "Array of message database IDs",
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "4oRBnPtCXgAqZniRhzLNmS",
                    "hXayS6wnCgNnt6aFTvmOF6"
                  ]
                },
                "Tags": {
                  "description": "Array of tag names to set",
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "Tag 1",
                    "Tag 2"
                  ]
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/OKResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/tags/{Tag}": {
      "put": {
        "description": "Renames an existing tag.",
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tags"
        ],
        "summary": "Rename a tag",
        "operationId": "RenameTagParams",
        "parameters": [
          {
            "type": "string",
            "description": "The url-encoded tag name to rename",
            "name": "Tag",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "Name"
              ],
              "properties": {
                "Name": {
                  "description": "New name",
                  "type": "string",
                  "example": "New name"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/OKResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "description": "Deletes a tag. This will not delete any messages with the tag, but will remove the tag from any messages containing the tag.",
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tags"
        ],
        "summary": "Delete a tag",
        "operationId": "DeleteTagParams",
        "parameters": [
          {
            "type": "string",
            "description": "The url-encoded tag name to delete",
            "name": "Tag",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/OKResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/view/{ID}.html": {
      "get": {
        "description": "Renders just the message's HTML part which can be used for UI integration testing.\nAttached inline images are modified to link to the API provided they exist.\nNote that is the message does not contain a HTML part then an 404 error is returned.\n\nThe ID can be set to `latest` to return the latest message.",
        "produces": [
          "text/html"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "testing"
        ],
        "summary": "Render message HTML part",
        "operationId": "GetMessageHTMLParams",
        "parameters": [
          {
            "type": "string",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Embed",
            "description": "If this is route is to be embedded in an iframe, set embed to `1` in the URL to add `target=\"_blank\"` and `rel=\"noreferrer noopener\"` to all links.\n\nIn addition, a small script will be added to the end of the document to post (postMessage()) the height of the document back to the parent window for optional iframe height resizing.\n\nNote that this will also *transform* the message into a full HTML document (if it isn't already), so this option is useful for viewing but not programmatic testing.",
            "name": "embed",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/HTMLResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    },
    "/view/{ID}.txt": {
      "get": {
        "description": "Renders just the message's text part which can be used for UI integration testing.\n\nThe ID can be set to `latest` to return the latest message.",
        "produces": [
          "text/plain"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "testing"
        ],
        "summary": "Render message text part",
        "operationId": "GetMessageTextParams",
        "parameters": [
          {
            "type": "string",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TextResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/NotFoundResponse"
          }
        }
      }
    }
  },
  "definitions": {
...
    "SMSMessage": {
      "description": "SMSMessage represents a stored SMS message",
      "type": "object",
      "properties": {
        "AccountSID": {
          "description": "Twilio AccountSID used to send the message",
          "type": "string"
        },
        "Body": {
          "description": "Message body",
          "type": "string"
        },
        "Created": {
          "description": "Received date \u0026 time",
          "type": "string",
          "format": "date-time"
        },
        "From": {
          "description": "Sender phone number",
          "type": "string"
        },
        "ID": {
          "description": "Database ID",
          "type": "string"
        },
        "Read": {
          "description": "Read status",
          "type": "boolean"
        },
        "To": {
          "description": "Recipient phone number",
          "type": "string"
        }
      },
      "x-go-package": "github.com/coreydaley/messagepit/internal/storage"
    },
    "SMSMessageSummary": {
      "description": "SMSMessageSummary is used for list views and websocket broadcasts",
      "type": "object",
      "properties": {
        "Body": {
          "description": "Message body (full — SMS messages are short)",
          "type": "string"
        },
        "Created": {
          "description": "Received date \u0026 time",
          "type": "string",
          "format": "date-time"
        },
        "From": {
          "description": "Sender phone number",
          "type": "string"
        },
        "ID": {
          "description": "Database ID",
          "type": "string"
        },
        "Read": {
          "description": "Read status",
          "type": "boolean"
        },
        "To": {
          "description": "Recipient phone number",
          "type": "string"
        }
      },
      "x-go-package": "github.com/coreydaley/messagepit/internal/storage"
    },
    "SMSMessagesSummary": {
      "type": "object",
      "title": "SMSMessagesSummary is the response shape for the SMS message list endpoint.",
      "properties": {
        "messages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SMSMessageSummary"
          },
          "x-go-name": "Messages"
        },
        "start": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Start"
        },
        "total": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Total"
        },
        "unread": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Unread"
        }
      },
      "x-go-package": "github.com/coreydaley/messagepit/server/apiv1"
    },
...
  },
  "responses": {
...
    "ArrayResponse": {
      "description": "Plain JSON array response",
      "schema": {
        "type": "array",
        "items": {
          "type": "string"
        }
      }
    },
...
    "ErrorResponse": {
      "description": "Server error will return with a 400 status code\nwith the error message in the body",
      "schema": {
        "type": "string"
      }
    },
    "HTMLResponse": {
      "description": "HTML response",
      "schema": {
        "type": "string"
      }
    },
...
    "NotFoundResponse": {
      "description": "Not found error will return a 404 status code",
      "schema": {
        "type": "string"
      }
    },
    "OKResponse": {
      "description": "Plain text \"ok\" response",
      "schema": {
        "type": "string"
      }
    },
    "SMSMessageResponse": {
      "description": "Single SMS message",
      "schema": {
        "$ref": "#/definitions/SMSMessage"
      }
    },
    "SMSMessagesSummaryResponse": {
      "description": "Summary of SMS messages",
      "schema": {
        "$ref": "#/definitions/SMSMessagesSummary"
      }
    },
...
    "TextResponse": {
      "description": "Plain text response",
      "schema": {
        "type": "string"
      }
    },
...
  }
}